requirements.txt
fastapi
uvicorn
groq
pinecone-client
sentence-transformers
python-dotenv
